add authentication and authorization by jwt

+++ controllers/api/v1/post_api.js
@@ -24,25 +24,18 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
-        // if (post.user == req.user.id) {
-        post.remove();
-        await Comment.deleteMany({ post: req.params.id });
-        return res.json(200, {
-            message: "post and associated comments deleted successfully"
-        });
-        // if (req.xhr) {
-        //     return res.status(200).json({
-        //         data: {
-        //             post: req.params.id
-        //         },
-        //         message: 'post deleted'
-        //     });
-        // }
-        // req.flash('success', 'post and associated comment is deleted');
-        // return res.redirect('back');
-        // } else {
-        //     return res.redirect('back');
-        // }
+        if (post.user == req.user.id) {
+            post.remove();
+            await Comment.deleteMany({ post: req.params.id });
+            return res.json(200, {
+                message: "post and associated comments deleted successfully"
+            });
+
+        } else {
+            return res.json(401, {
+                message: 'You cannot delete this post!'
+            });
+        }
 
 
     }
